recipetool: create: avoid extra blank lines in output recipe
If we output extra blank lines (because of some automated editing) then
it makes the output recipe look a bit untidy. You could argue that we
should simply have the editing code not do that, but sometimes we don't
have enough context there to do that. It's simple enough to just filter
out the extra blank lines when writing the file, so just do it that way.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>