kernel-yocto: allow kconfig mode to be specified for SRC_URI defconfigs
It isn't possible to have defaults that meet all the processing requirements
of defconfigs. i.e. we don't know if they are complete, or partial. We don't
know if they are from an older kernel, or the current one. We don't know if
defaults for new symbols are desired, or if any non specified options should
be disabled.

The KCONFIG_MODE variable was added to add flexibility in the processing to
allow a recipe author to add explicit direction. That being said, there were
still use cases that weren't being handled as nicely as they could.

To enable the upgrade and savedefconfig use cases for SRC_URI based defconfig
elements, we introduce the processing of a "kconfig_mode" parameter.
This allows any fragments or defconfig on the SRC_URI to specify the
kconfig mode: Valid modes are "alldefconfig" and "allnoconfig"

In the case of competing modes, the last fragment (or defconfig) wins

[YOCTO #8743]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>