scripts/runqemu: Be more loose when guessing QB_SYSTEM_NAME variable
Following the instructions to reproduce YOCTO #10026 I found that
runqemu would fail to start because QB_SYSTEM_NAME is null. This patch
makes the guessing algorithm more loose allowing it to correctly guess
the variable from the MACHINE name.

It is still a good assumption that if it contains "qemux86-64" for
example, that the system is x86-64, so this is unlikely to cause any
new issues while preserving some peoples' old workflows.

V2 changes:
* Fix if/elif block to actually start with if

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>