devtool: add rename subcommand
When you run devtool add on a source tree we attempt to figure out the
correct name and version for the recipe. However, despite our best
efforts, sometimes the name and/or version we come up with isn't
correct, and the only way to remedy that was to reset the recipe, delete
the source tree and start again, specifying the name this time. To avoid
this slightly painful procedure, add a "rename" subcommand that lets you
rename the recipe and/or change the version.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>