summary: change format to be a bit more readable
In order to place the emphasis on information that is most important to
the submitter, we need to alter the presentation in the default jinja2
template to present the commit first (if applicable), the issue (suffixed
with the test name), the suggested fix (if any), and then any optional
additional fields.

In order to do this and at the same time allow the template a bit more
control over formatting, require tests to send a dict with "id" and
"issue" elements, plus optionally "fix", "commit" and "data". These are
passed verbatim into the jinja2 template when formatting the summary.
(As a general rule though for human readability I believe we should be
expressing additional information as part of the issue / fix text rather
than supplying it in field form via the data parameter.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>