summary: fix parameter passing
At the moment we're referring to a module-level variable ("opt") in the
function, and we were passing certain parameters into the rendering that
aren't necessary (template name, output file name). To fix both of
these, pass some of these explicitly to the function and then pass the
parameters to go into the rendering separately.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>