data_smart: Default to expansion for getVar/getVarFlags
We've been building to this for a while, default to return expanded
values for getVar/getVarFlags.

We can then go through and remove the "True" option to many of the
calls to this function, all funcion calls should have a default by now
though since the parameter has been required for a while.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>