use multiple processes to dump signatures.
This change significantly shortens the time at reparsing
stage of '-S' option.

It fires multiple processes in batches. Each file is reparsed
and then dumped within a dedicated process. The maximum number
of the running processes is not greater than the value of
BB_NUMBER_PARSE_THREADS if it is set.

The dump_sigs() in class SignatureGeneratorBasic is _replaced_
by a new dump_sigfn() interface, so calls from the outside and
subclasses are dispatched to the implementation in the base
class of SignatureGeneratorBasic

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>