Using the computed when basehash mismatch
I am not sure if the current code behaves as it's expected. It seems
in the dump_sigtask() the old basehash will be used to calculate
taskhash when the new computed basehash mismatches the old.

This patch passes the new computed basehash to calc_taskhash() in this
corner case, but doesn't update basehash in sig generator's instance.

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>