fetch2/npm: fix handling of os field
When I wrote this check I didn't quite understand how the values in this
field should be expressed - it seems from reading the documentation if
there is an entry starting with '!' then the list is a blacklist and we
shouldn't expect "linux" to be in the list, or we'll end up skipping
important dependencies.

This fixes fetching the "statsd" npm package.

Fixes [YOCTO #10760].

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>