oeqa/core/context.py: Adds compatibility with bitbake logger
The bitbake logger changes the way debug is logged and adds
different levels within debug, this is passed as argument
to the function and breaks compatibility with vanilla loggers.

This implements a way to handle this adding a new function for
debug, that will dispatch the correct logging method depending
on received arguments. It is not pretty, but allows to have
bitbake and vainilla loggers.

Also this commit fix the issue of not showing the test summary
and results when running from bitbake.

[YOCTO #10686]

Signed-off-by: Mariano Lopez <mariano.lopez@linux.intel.com>