gettext_0.19.8.1.bb: Add 'gettext-libintl' package
Whilst gettext is not the default provider of libintl for most libc
configurations it is possible for it to be the provider of libintl. In
the cases where it is provided it should be shipped in a separate
package so as to allow for it to be included without pulling in
additional gettext content/dependencies.

This change adds a package called 'gettext-libintl' which ships the
libintl shared objects (and is otherwise empty if libintl is not
provided). This allows for the library to be shipped without the
additional components included in the gettext-runtime package.

This package is prefixed with 'gettext-' to define the source of the
library due to the existence of multiple providers for libintl.

This change also makes the libintl*.so and .h packaged into the
gettext-runtime-dev package by default. Removing the need for a uclibc
override to handle packaging these files. Targets that don't intend to
distribute these files already remove them during the do_install stage
(e.g. musl).

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>